Notification box added to bottom of screen, zoom fixed + border + other changes
Graphic Raycast script added to the Canvas to detect when user touches notificationbox.
This function produces a bool which is used to begin the panning/zooming functions or not.
If the raycast detects the notification box, then the map will not pan/zoom.
This notification box will be used to display notifications about the empire and the current climate.
Zoom function fixed, retro border added to HUD.